% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{data}
\alias{data}
\title{Atlas of Hillforts of Britain and Ireland data plus shapefile}
\format{One object of class \code{"Large SpatialPolygonsDataFrame"} called "shape"; two objects  of class \code{"Dataframe"} called "data" and "distance_table"}
\usage{
data(data)
}
\description{
The example point coordinate data is taken from the:
Atlas of Hillforts of Britain and Ireland 
Lock, G. and Ralston, I. 2017.  Atlas of Hillforts of Britain and Ireland. [ONLINE] Available at: https://hillforts.arch.ox.ac.uk
This work is licensed under a Creative Commons Attribution-ShareAlike 4.0 International License
The point coordinate data was extracted from the database for hillforts in Britain (England, Scotland, Wales and the Isle of Man) which have status of interpretation as "confirmed". This was extracted on: 25/07/2017
The example shape file is an outline of Britain, taken from: 
Ordnance Survey OS Open Data, uner a UK Open Government Licence: https://www.ordnancesurvey.co.uk/opendatadownload/products.html#OVERGB
This file Contains OS data © Crown Copyright (2020)
For example of usage, please see the vignette.
}
\references{
Atlas of Hillforts of Britain and Ireland, Lock, G. and Ralston, I. 2017.  Atlas of Hillforts of Britain and Ireland. [ONLINE] Available at: https://hillforts.arch.ox.ac.uk

Ordnance Survey OS Open Data, uner a UK Open Government Licence: https://www.ordnancesurvey.co.uk/opendatadownload/products.html#OVERGB
}
\keyword{datasets}
