% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{hillforts}
\alias{hillforts}
\title{Atlas of Hillforts of Britain and Ireland data}
\format{Object of class \code{"Dataframe"} called "hillforts"}
\usage{
data(data)
}
\description{
The example point coordinate data is taken from the:
Atlas of Hillforts of Britain and Ireland 
Lock, G. and Ralston, I. 2017.  Atlas of Hillforts of Britain and Ireland. [ONLINE] Available at: https://hillforts.arch.ox.ac.uk
This work is licensed under a Creative Commons Attribution-ShareAlike 4.0 International License
The point coordinate data was extracted from the database for hillforts in Britain (England, Scotland, Wales and the Isle of Man) which have status of interpretation as "confirmed". This was extracted on: 25/07/2017.
For example of usage, please see the vignette "percolation".
}
\references{
Atlas of Hillforts of Britain and Ireland, Lock, G. and Ralston, I. 2017.  Atlas of Hillforts of Britain and Ireland. [ONLINE] Available at: https://hillforts.arch.ox.ac.uk
}
\keyword{datasets}
