% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping_clusters.v4.12.R
\name{mapClusters}
\alias{mapClusters}
\title{CLUSTER MAPPING SCRIPT
Original code from Elsa Arcaute, CASA, UCL 28.2.15; extensively modified by Simon Maddison.
Cite as:	arXiv:1504.08318 [physics.soc-ph]
Version 3a: 10/5/15 & 2/10/15 - with additions to include county boundary outlines,
This code maps the clusters overlaid on an outline of the British Isles or Ireland including county outlines
England and Wales shape files added from Portsmouth University files 31/1/16
Version 4: 24/3/16. Changes to plots made to:
	Include identity of clusters (i.e.cluster index) in legend - done
	Plot the source of the original data (e.g. Atlas, Hogg etc.) (taken from source file name)- done
	Parameterise the shape file to be used for plotting ...
Take out graded grey colours for clusters > 15
Version 4.1: 29/3/16 changed outputs to be multi-page pdf
NOTE: Ireland not plotting points TO BE RESOLVED
Version 4.2: 6/4/16 - error in plot of all site corrected; removed 'xy'; all refs now to xy_data
Version 4.3: 10/4/16 - output shape file with all data points, with same projection as input map
    NOTE: esri shape file driver truncates names to 10 characters but this should not be a
    problem at the moment. Prob of e.g. PlacesIndex, ClusterNumber etc.
	shape file name is name of source file without .csv etc. extension
Version 4.4: 29/4/16 - paging upgraded to export to A4 and use the full page
Version 4.5: 6/5/16 - changed Index name to fit within 8 chars -> PlcIndex
Version 4.6: 8/5/16 - working on Ireland plot to fix. Problem seemed to be an old Ireland map with
an obsolete coordinate system
Add in external input to country and shape file for source,
file is added to source_file.txt - map name and shape file name, headers also added
source_file, map_name, shape_file
REQUIRES CHANGE TO ALL OTHER PROGRAMS (done)
Version 4.7: 18/5/16 -  Cluster data now comes in a single file (member_cluster_by_radius.csv)
  	output cluster ranks as new file site_cluster_rank_by_radius, lists by PlcIndex the cluster rank
indexed by radius
Version 4.8: 30/7/16 - Plot as png file output. Earlier issues with resolution resolved.
Easier to include in word document than pdf. However each plot is a separate file,
with suitable name increment
Version 4.9: 03/08/2016 - minor change to put legend on right for Domesday counties
Version 4.10: 7/11/2017 - working directory amended for new pc
Tweaks to legend location, size and plot title to allow for other data sets
Version 4.11: 24/06/2018 - only worked for integer increments of radius. Modified to allow fractional
values of radius, e.g. step size of 0.1km
Version 4.12: 04/12/18: - add in unit parameter to select km or m, in "radius_values.txt"
NOTE: values are assumed to be in metres, unit will factor this in calculations and charts by the unit value
so that a unit value of 1 shows through as metres, and a value of 1000 as km. Other values accomodated
Also modified so that when incrementing in decimal step values, all map names are formatted consistently}
\usage{
mapClusters(x)
}
\description{
CLUSTER MAPPING SCRIPT
Original code from Elsa Arcaute, CASA, UCL 28.2.15; extensively modified by Simon Maddison.
Cite as:	arXiv:1504.08318 [physics.soc-ph]
Version 3a: 10/5/15 & 2/10/15 - with additions to include county boundary outlines,
This code maps the clusters overlaid on an outline of the British Isles or Ireland including county outlines
England and Wales shape files added from Portsmouth University files 31/1/16
Version 4: 24/3/16. Changes to plots made to:
	Include identity of clusters (i.e.cluster index) in legend - done
	Plot the source of the original data (e.g. Atlas, Hogg etc.) (taken from source file name)- done
	Parameterise the shape file to be used for plotting ...
Take out graded grey colours for clusters > 15
Version 4.1: 29/3/16 changed outputs to be multi-page pdf
NOTE: Ireland not plotting points TO BE RESOLVED
Version 4.2: 6/4/16 - error in plot of all site corrected; removed 'xy'; all refs now to xy_data
Version 4.3: 10/4/16 - output shape file with all data points, with same projection as input map
    NOTE: esri shape file driver truncates names to 10 characters but this should not be a
    problem at the moment. Prob of e.g. PlacesIndex, ClusterNumber etc.
	shape file name is name of source file without .csv etc. extension
Version 4.4: 29/4/16 - paging upgraded to export to A4 and use the full page
Version 4.5: 6/5/16 - changed Index name to fit within 8 chars -> PlcIndex
Version 4.6: 8/5/16 - working on Ireland plot to fix. Problem seemed to be an old Ireland map with
an obsolete coordinate system
Add in external input to country and shape file for source,
file is added to source_file.txt - map name and shape file name, headers also added
source_file, map_name, shape_file
REQUIRES CHANGE TO ALL OTHER PROGRAMS (done)
Version 4.7: 18/5/16 -  Cluster data now comes in a single file (member_cluster_by_radius.csv)
  	output cluster ranks as new file site_cluster_rank_by_radius, lists by PlcIndex the cluster rank
indexed by radius
Version 4.8: 30/7/16 - Plot as png file output. Earlier issues with resolution resolved.
Easier to include in word document than pdf. However each plot is a separate file,
with suitable name increment
Version 4.9: 03/08/2016 - minor change to put legend on right for Domesday counties
Version 4.10: 7/11/2017 - working directory amended for new pc
Tweaks to legend location, size and plot title to allow for other data sets
Version 4.11: 24/06/2018 - only worked for integer increments of radius. Modified to allow fractional
values of radius, e.g. step size of 0.1km
Version 4.12: 04/12/18: - add in unit parameter to select km or m, in "radius_values.txt"
NOTE: values are assumed to be in metres, unit will factor this in calculations and charts by the unit value
so that a unit value of 1 shows through as metres, and a value of 1000 as km. Other values accomodated
Also modified so that when incrementing in decimal step values, all map names are formatted consistently
}
