% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering_script.v3.6.R
\name{percolate}
\alias{percolate}
\title{CLUSTERING SCRIPT, RUNS AFTER NODES LIST SCRIPT
Original code from Elsa Arcaute, CASA, UCL 28.2.15; extensively modified by Simon Maddison.
Cite as:	arXiv:1504.08318 [physics.soc-ph]
Version 3
This is version 3, incorporating adjustments and enhancements to code 10/5/15 & 2/10/15
This code computes the clusters, based on matrix of inter-nodal distances computed in script
 	create_nodes_list_d.vxxx.R
Note that the list of nodes and internodal distances is not a complete matrix, but only includes
	distances below a certain threshold. This keeps the data sizes manageable.
Note that the original used library bigmemory but this is not available for windows so all refs removed.
Note that when nodes are individual, they are no longer in a cluster. (31/1/16)
Version 3.1: 20/04/16 - Note that this intrinsically cannot generate clusters with only one node.
  	Need to post process means in the analysis script
Version 3.2: 6/5/16 - Changed Index name to fit within 8 chars -> PlcIndex
Version 3.3: 18/5/16 - output data into a single file, indexed on PlcIndex and radius instead of separate
files. Read in PlcIndex.csv generated by create_nodes_list_d....R
Version 3.4: 7/11/17 - path of working directory amended for new PC
Version 3.6: 04/12/18 - changes in Version 3.5 omitted
	Add in unit parameter to select km or m, in "radius_values.txt"
NOTE: values are assumed to be in metres, unit will factor this in calculations and charts by the unit value
so that a unit value of 1 shows through as metres, and a value of 1000 as km. Other values accomodated.
	NOTE: no modification necessary to work for fractional values of radius, e.g. step size of 0.1km???????????????}
\usage{
percolate(x)
}
\description{
CLUSTERING SCRIPT, RUNS AFTER NODES LIST SCRIPT
Original code from Elsa Arcaute, CASA, UCL 28.2.15; extensively modified by Simon Maddison.
Cite as:	arXiv:1504.08318 [physics.soc-ph]
Version 3
This is version 3, incorporating adjustments and enhancements to code 10/5/15 & 2/10/15
This code computes the clusters, based on matrix of inter-nodal distances computed in script
 	create_nodes_list_d.vxxx.R
Note that the list of nodes and internodal distances is not a complete matrix, but only includes
	distances below a certain threshold. This keeps the data sizes manageable.
Note that the original used library bigmemory but this is not available for windows so all refs removed.
Note that when nodes are individual, they are no longer in a cluster. (31/1/16)
Version 3.1: 20/04/16 - Note that this intrinsically cannot generate clusters with only one node.
  	Need to post process means in the analysis script
Version 3.2: 6/5/16 - Changed Index name to fit within 8 chars -> PlcIndex
Version 3.3: 18/5/16 - output data into a single file, indexed on PlcIndex and radius instead of separate
files. Read in PlcIndex.csv generated by create_nodes_list_d....R
Version 3.4: 7/11/17 - path of working directory amended for new PC
Version 3.6: 04/12/18 - changes in Version 3.5 omitted
	Add in unit parameter to select km or m, in "radius_values.txt"
NOTE: values are assumed to be in metres, unit will factor this in calculations and charts by the unit value
so that a unit value of 1 shows through as metres, and a value of 1000 as km. Other values accomodated.
	NOTE: no modification necessary to work for fractional values of radius, e.g. step size of 0.1km???????????????
}
