% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_clust_freq.R
\name{percolate_all}
\alias{percolate_all}
\title{created 2019-08-01 by Sophie Schmidt}
\usage{
percolate_all(data, radius_values, limit, radius_unit, upper_radius,
  lower_radius, step_value)
}
\description{
This is a function built based on the scripts create_nodes_list_d.v4.7, clustering_script.v.3.6 and cluster_frequency_script.v2.17 
by Simon Maddison
it creates a file structure in which working data and analysis results are saved
input data needed, which consists of a dataframe with the PLCINDEX, EASTING, NORTHING - header
input of radius values needed
output: in working_data: csv with the radius-input data, PlcIndex, null_entries and duplicate_entries
output in analysis_results: analysis_by_radius.csv and member_cluster_by_radius.csv
both output-folders will be used by following functions
For computation time
List any entries with null values and write to file
Remove rows with null values from data
Remove points that are superimposed and keep only the first ID
- This removes one of two sites that are very close to each other
Determined on basis of x y coordinates
Write list to file
Write file with list of PlcIndex used
Number of points/nodes in file with no duplicates
Create matrix of internodal distances
Columns: NodeId1, NodeId2, distance 1-2
Compute distance between nodes - pythagoras
 for full grid references this is in units of 1m
to give distance in m*unit, rounded to 2 decimal places; this also reduces file size
Include only if less than limit of distances to be included
Output file name and location writes to a text file
Remove the unused rows in the matrix
need to write this WITHOUT the row number
}
