% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percopackage.R
\docType{package}
\name{percopackage}
\alias{percopackage}
\alias{percopackage-package}
\title{\code{percopackage} package}
\description{
percolation analysis in R
}
\details{
copied from Jenny Brian
\href{https://github.com/SCSchmidt/percopackage/README.html}{GitHub}
}
