% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_freq.R
\name{plot_clust_freq}
\alias{plot_clust_freq}
\title{CLUSTER FREQENCY PLOTTING SCRIPT
Simon Maddison, 10th May 2015
Version 2.1: 2/10/15 - Tidied up, paths and added plot of radius vs #' of clusters
Version 2.2: 31/1/16 -  Labels added to the second plot, requires the textxy function in calibrate package
Note odd plot of radius 1 if included, due to clustering idiosyncracy that does not include clusters of 1.
Version 2.3: 30/3/16 - Extended to include plots of average and median cluster sizes on the same plot
	(currently max size). Changed outputs to multi-page pdf files
Version 2.4: 20/4/16 - Plot to log log added for median. Defined range plot removed
Version 2.5: 23/4/16 -
Added in computations for clusters of single nodes
	Totals of nodes and number of clusters take account of single nodes as clusters of 1
Table of max, mean and median exported
Version 2.6: 26/4/16 -
	As per Elsa's paper, plot 'S_mean' cluster size, vs radius,
	Exclude clusters with less than a certain threshold of nodes (parameter)
	Also exclude largest cluster from mean
	Thresholds for upper and lower calculations of S_mean read in from radius_values text file
Version 2.7: 29/4/16 -
As per Elsa's published paper, plot max cluster size vs. radius as well
	Page format improved to generate pdf A4 format
	Bug fixed when adding clusters of size 1 (if null value)
Version 2.8: 4/5/16 -
As per Elsa's published paper, max also plotted as normalized relative to total number of nodes
Version 2.9: 6/5/16 -
Problem with S_mean if number of clusters becomes zero, resolved, and also corrected computation
Version 2.10: 8/5/16 - source text file format changed to include map information
Version 2.11: 18/5/16 - updated to read cluster data from combined file that is now generated by other programs
File is member_cluster_by_radius.csv
Version 2.12: 30/7/16 - data output as png file, easier to incorporate in word doc
Make the image square or else it distorts the plot (note that this does not happen in pdf)
	Remove simple cluster rank plots, no longer useful
Version 2.13: 7/11/17	- amended working directory path for new pc
Version 2.14: 17/11/17 - amended to allow for multiple file versions based on additional 'k' factor
used by dbScan. For each run there may be more than one value of k, so one file generated per value
	value included in the file name. NOTE THIS WILL NEED AMENDING FOR PERCOLATION SCRIPTS TO UPDATE
Some plots removed, legacy of earlier attempts at analysis
Version 2.15: 10/01/18 - if dbscan values are null, then treat as percolation and name files accordingly.
Version 2.16: 03/12/18 - Add in unit parameter to select km or m, in "radius_values.txt"
NOTE: values are assumed to be in metres, unit will factor this in calculations and charts by the unit value
so that a unit value of 1 shows through as metres, and a value of 1000 as km. Other values accomodated.
	Modified to allow fractional values of radius, e.g. step size of 0.1km
S mean computations removed.
Version 2.17: 02/04/2019 - additional variants of plots generated; without labels and a continuous line plot;
add plot of cluster frequency vs. percolation radius. Added to use in presentations and publications
	Minor ticks added to axes on certain plots
Background set to transparent, for overlay plots; line weight increased
Removed inclusion of clusters size 1, i.e. nodes not inclusters
This script extracts cluster frequency data and plots it for a range of percolation radii
The data is in the working directory in text files;
	each file is a list of Place indices, with the identity of each cluster
	which it is a member of.}
\usage{
plot_clust_freq(source_file_name)
}
\description{
CLUSTER FREQENCY PLOTTING SCRIPT
Simon Maddison, 10th May 2015
Version 2.1: 2/10/15 - Tidied up, paths and added plot of radius vs #' of clusters
Version 2.2: 31/1/16 -  Labels added to the second plot, requires the textxy function in calibrate package
Note odd plot of radius 1 if included, due to clustering idiosyncracy that does not include clusters of 1.
Version 2.3: 30/3/16 - Extended to include plots of average and median cluster sizes on the same plot
	(currently max size). Changed outputs to multi-page pdf files
Version 2.4: 20/4/16 - Plot to log log added for median. Defined range plot removed
Version 2.5: 23/4/16 -
Added in computations for clusters of single nodes
	Totals of nodes and number of clusters take account of single nodes as clusters of 1
Table of max, mean and median exported
Version 2.6: 26/4/16 -
	As per Elsa's paper, plot 'S_mean' cluster size, vs radius,
	Exclude clusters with less than a certain threshold of nodes (parameter)
	Also exclude largest cluster from mean
	Thresholds for upper and lower calculations of S_mean read in from radius_values text file
Version 2.7: 29/4/16 -
As per Elsa's published paper, plot max cluster size vs. radius as well
	Page format improved to generate pdf A4 format
	Bug fixed when adding clusters of size 1 (if null value)
Version 2.8: 4/5/16 -
As per Elsa's published paper, max also plotted as normalized relative to total number of nodes
Version 2.9: 6/5/16 -
Problem with S_mean if number of clusters becomes zero, resolved, and also corrected computation
Version 2.10: 8/5/16 - source text file format changed to include map information
Version 2.11: 18/5/16 - updated to read cluster data from combined file that is now generated by other programs
File is member_cluster_by_radius.csv
Version 2.12: 30/7/16 - data output as png file, easier to incorporate in word doc
Make the image square or else it distorts the plot (note that this does not happen in pdf)
	Remove simple cluster rank plots, no longer useful
Version 2.13: 7/11/17	- amended working directory path for new pc
Version 2.14: 17/11/17 - amended to allow for multiple file versions based on additional 'k' factor
used by dbScan. For each run there may be more than one value of k, so one file generated per value
	value included in the file name. NOTE THIS WILL NEED AMENDING FOR PERCOLATION SCRIPTS TO UPDATE
Some plots removed, legacy of earlier attempts at analysis
Version 2.15: 10/01/18 - if dbscan values are null, then treat as percolation and name files accordingly.
Version 2.16: 03/12/18 - Add in unit parameter to select km or m, in "radius_values.txt"
NOTE: values are assumed to be in metres, unit will factor this in calculations and charts by the unit value
so that a unit value of 1 shows through as metres, and a value of 1000 as km. Other values accomodated.
	Modified to allow fractional values of radius, e.g. step size of 0.1km
S mean computations removed.
Version 2.17: 02/04/2019 - additional variants of plots generated; without labels and a continuous line plot;
add plot of cluster frequency vs. percolation radius. Added to use in presentations and publications
	Minor ticks added to axes on certain plots
Background set to transparent, for overlay plots; line weight increased
Removed inclusion of clusters size 1, i.e. nodes not inclusters
This script extracts cluster frequency data and plots it for a range of percolation radii
The data is in the working directory in text files;
	each file is a list of Place indices, with the identity of each cluster
	which it is a member of.
}
