% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClustFreq.v2.2.R
\name{plotClustFreq}
\alias{plotClustFreq}
\title{Plot Cluster Frequency}
\usage{
plotClustFreq(source_file_name = source_file_name)
}
\arguments{
\item{source_file_name}{needs input of source file name for plotting}
}
\value{
three plots of radius to maximum cluster size, radius to 
 mean cluster size and radius to normalized max. cluster size
}
\description{
Percolation cluster size and rank analyses 
– this program processes the cluster data generated by the cluster extraction program to give various different frequency and ranking plots.
}
\details{
source_file_name is a string, which will be used to label the plots and describes the source data set. 
If already input in mapClusters this doesn’t need to be done again. 

Three png-files are generated:

a) radius to maximum cluster size, 

b) radius to mean cluster size and 

c) radius to normalized max. cluster size
and they will include the information of the source file as a subtitle.

They are stored as png files in the /analysis_results directory as “radius_to_max_cluster_size”, “radius_to_mean_cluster_size” and “radius_to_norm_max_cluster_size”.

For more information and a code and data example please check the vignette "percolation".
}
\author{
Simon Maddison

Sophie C. Schmidt
}
