% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_shape.R
\docType{data}
\name{GB_shape}
\alias{GB_shape}
\title{Shapefile: Outline of Great Britain}
\format{
Object of class \code{"Large SpatialPolygonsDataFrame"} called "GB_shape"
}
\usage{
data(data)
}
\description{
Ordnance Survey OS Open Data, under a UK Open Government Licence: https://www.ordnancesurvey.co.uk/opendatadownload/products.html#OVERGB
This file Contains OS data © Crown Copyright (2020)
}
\details{
For example of usage, please see the vignette "percolation".
}
\references{
Ordnance Survey OS Open Data, uner a UK Open Government Licence: https://www.ordnancesurvey.co.uk/opendatadownload/products.html#OVERGB
}
\keyword{datasets}
