% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_clust.R
\name{mapClusters}
\alias{mapClusters}
\title{created 2019-08-01 by Sophie Schmidt}
\usage{
mapClusters(data, shape, map_name)
}
\arguments{
\item{data}{needs input of dataframe}

\item{shape}{needs input of SpatialPolygonDataFrame from 'sp'}

\item{map_name}{needs string for how the maps shall be named}
}
\value{
as many maps as png as step_values have been calculated
}
\description{
This is a function built based on the script mapping_clusters.v4.12
by Simon Maddison
it creates a folder maps, where output is saved as png
}
