% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapClusters.v.2.1.R
\name{mapClusters}
\alias{mapClusters}
\title{mapClusters}
\usage{
mapClusters(shape, map_name, source_file_name, dpi = 300)
}
\arguments{
\item{shape}{needs input of SpatialPolygonDataFrame from 'sp' for background map (same ESPG!)}

\item{map_name}{needs string for how the maps shall be named}

\item{source_file_name}{string for the name of data set}

\item{dpi}{dots per inch for the images. set to 300 as standard}
}
\value{
as many maps as png as step_values have been calculated
}
\description{
This is a function built based on the script mapping_clusters.v4.12
by Simon Maddison
it creates a folder maps, where output is saved as png
}
\author{
Simon Maddison

Sophie C. Schmidt
}
