% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapClusters.v.2.2.R
\name{mapClusters}
\alias{mapClusters}
\title{Map clusters}
\usage{
mapClusters(shape, map_name, source_file_name, dpi = 300)
}
\arguments{
\item{shape}{needs input of SpatialPolygonDataFrame from 'sp' for background map (same ESPG!)}

\item{map_name}{needs string for how the maps shall be named}

\item{source_file_name}{string for the name of data set used, to include on maps}

\item{dpi}{dots per inch for the images. set to 300 as standard}
}
\value{
as many maps as png as step_values have been calculated
}
\description{
This function will map the clusters created by percolation.
}
\details{
shape is a SpatialPolygonDataFrame (sp), which will be used for plotting as a background map, and determining the coordinate system.

map_name is a string, which will be used to title the maps.

source_file_name is a string, which will be used in a subtitle within the maps and should describe the source data used.

dpi is the output file resolution. The default is set to 300. Other resolutions may be needed for publication, for example.

Exported maps are in the folder "/maps": As many png map files as there are different radii values used in the percolation analysis. Also a shp-file with the point data.

For more information and a code and data example please check the vignette.
}
\author{
Simon Maddison

Sophie C. Schmidt
}
