% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nodes_clust_freq.R
\name{percolate_all}
\alias{percolate_all}
\title{created 2019-08-01 by Sophie Schmidt}
\usage{
percolate_all(data, limit, radius_unit, upper_radius, lower_radius,
  step_value)
}
\arguments{
\item{data}{needs input of dataframe}

\item{limit}{needs input of integer: is the value above which distances will not be calculated between sites}

\item{radius_unit}{is either 1 for meter or 1000 for km for all input numbers}

\item{upper_radius}{needs input of integer, is the upper value of the radius range to be used}

\item{lower_radius}{needs input of integer, is the lower value of the radius range to be used}

\item{step_value}{integer or numeric, is the step value to be used between these two values}
}
\value{
3 plots of radius to maximum cluster size, radius to mean cluster size and radius to normalized max. cluster size
}
\description{
This is a function built based on the scripts create_nodes_list_d.v4.7, clustering_script.v.3.6 and cluster_frequency_script.v2.17 
by Simon Maddison
it creates a file structure in which working data and analysis results are saved
input data needed, which consists of a dataframe with the PLCINDEX, EASTING, NORTHING - header
input of radius values needed
output: in working_data: csv with the radius-input data, PlcIndex, null_entries and duplicate_entries
output in analysis_results: analysis_by_radius.csv and member_cluster_by_radius.csv
both output-folders will be used by following functions
}
